fix new

+++ elboutique/src/app/wish-list/wish-list-content/wish-list-content.component.ts
@@ -13,6 +13,7 @@ import { RouterLink } from '@angular/router';
 import { CartService } from '../../service/cart.service';
 import { faTrashCan } from '@fortawesome/free-solid-svg-icons';
 import { faHeart } from '@fortawesome/free-solid-svg-icons';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-wish-list-content',
   standalone: true,
@@ -65,18 +66,29 @@ export class WishListContentComponent implements OnInit, OnDestroy {
       });
     });
   }
+  // with the sweetalert 
   removeProductFromWishlist(id: number) {
-    const sentbody: Object = {
-      products: id,
-    };
-
-    this.wishlistService.deleteItemFromWishlist(sentbody);
-    this.toaster.error('Product Removed From Wishlist', 'Remove');
-    this.userWishlist = this.userWishlist.filter(
-      (product: any) => product.id != id
-    );
+    Swal.fire({
+      title:'Are you sure ?',
+      text:'Do you really want to delete this product from your wishlist ?',
+      icon:'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, remove it!',
+      cancelButtonText: 'No, keep it',
+    }).then((reult) =>{
+      if (reult.isConfirmed){
+      const sentbody: Object = {
+        products: id,
+      };
+  
+      this.wishlistService.deleteItemFromWishlist(sentbody);
+      this.toaster.error('Product Removed From Wishlist', 'Remove');
+      this.userWishlist = this.userWishlist.filter(
+        (product: any) => product.id != id
+      );
+      }
+    })
   }
-
   addProductToCart(e: HTMLButtonElement, id: number) {
     const sentbody: Object = {
       products: {
@@ -96,6 +108,7 @@ export class WishListContentComponent implements OnInit, OnDestroy {
     };
     this.cartService.deleteItemFromCart(sentbody);
     this.toaster.error('Product Removed From Cart', 'Remove');
+    console.log(sentbody);
     // Change button text
     e.textContent = 'Add to cart';
   }
